requirements.txt
streamlit
numpy
pandas
matplotlib
scipy
python-dotenv